add support python3

+++ lib/python/Components/TunerInfo.py
@@ -1,4 +1,6 @@
-from GUIComponent import GUIComponent
+#!/usr/bin/python
+# -*- coding: utf-8 -*-
+from Components.GUIComponent import GUIComponent
 
 from enigma import eLabel, eSlider, iFrontendInformation
 
@@ -45,7 +47,7 @@ class TunerInfo(GUIComponent):
 		if not val:
 			return 0
 		if val < 2500:
-			return long(log(val) / log(2))
+			return int(log(val) / log(2))
 		return val * 100 / 65535
 
 	def update(self):
